User Story
"As a user, I want to be able to filter cast search results by one facet, metacritic rating."

Task

For this ticket, you'll be required to modify the getMoviesCastFaceted method in MovieDao.java, so the MFlix application can perform faceted searches. You will find a more detailed description of the task as a comment in the MovieDao.java file.